Organizing the AbsenceGenerator
